+++ src/java-apps/gateway/src/main/java/in/zycon/demo/gateway/Gateway.java
@@ -1,0 +1,27 @@
+package in.zycon.demo.gateway;
+
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.context.annotation.Bean;
+import springfox.documentation.builders.PathSelectors;
+import springfox.documentation.builders.RequestHandlerSelectors;
+import springfox.documentation.spi.DocumentationType;
+import springfox.documentation.spring.web.plugins.Docket;
+
+@SpringBootApplication
+public class Gateway {
+
+    public static void main(String[] args) {
+        SpringApplication.run(Gateway.class, args);
+    }
+
+    @Bean
+    public Docket api() {
+        return new Docket(DocumentationType.SWAGGER_2)
+                .select()
+                .apis(RequestHandlerSelectors.basePackage("in.zycon.demo.gateway.controller"))
+                .paths(PathSelectors.any())
+                .build();
+    }
+}
