+++ src/java-apps/news/src/main/java/in/zycon/demo/gateway/controller/NewsAPI.java
@@ -1,0 +1,32 @@
+package in.zycon.demo.gateway.controller;
+
+import in.zycon.demo.gateway.model.News;
+import in.zycon.demo.gateway.service.NewsService;
+import io.swagger.annotations.ApiOperation;
+import io.swagger.annotations.ApiResponse;
+import io.swagger.annotations.ApiResponses;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+public class NewsAPI {
+    @Autowired
+    NewsService newsService;
+
+
+    @GetMapping(value = {"/news", "/news/{id}"})
+    @ApiOperation(value = "Get All News", response = News.class)
+    @ApiResponses(value = {
+            @ApiResponse(code = 200, message = "News Fetched successfully")
+    })
+    public ResponseEntity<?> getProduct(@PathVariable(required = false, name = "id") Integer id) {
+        if (id != null) {
+            return ResponseEntity.ok(newsService.getNewsById(id));
+        } else {
+            return ResponseEntity.ok(newsService.getAllNews());
+        }
+    }
+}
