+++ src/java-apps/news/src/main/java/in/zycon/demo/gateway/exception/NewsNotFound.java
@@ -1,0 +1,8 @@
+package in.zycon.demo.gateway.exception;
+
+public class NewsNotFound extends RuntimeException {
+
+    public NewsNotFound(String message) {
+        super(message);
+    }
+}
