+++ src/java-apps/news/src/main/java/in/zycon/demo/gateway/exception/PartnerDataException.java
@@ -1,0 +1,8 @@
+package in.zycon.demo.gateway.exception;
+
+public class PartnerDataException extends RuntimeException {
+
+    public PartnerDataException(String message) {
+        super(message);
+    }
+}
