+++ src/java-apps/news/src/main/java/in/zycon/demo/gateway/partner/Helper.java
@@ -1,0 +1,18 @@
+package in.zycon.demo.gateway.partner;
+
+import in.zycon.demo.gateway.properties.NewsPartner;
+import in.zycon.demo.gateway.properties.Partners;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Component;
+
+@Component
+public class Helper {
+    @Autowired
+    Partners partners;
+
+    public NewsPartner getPartnerByName(String name){
+        return partners.getNews().stream().filter(newsPartner -> {
+            return newsPartner.getName().equalsIgnoreCase(name);
+        }).findFirst().get();
+    }
+}
