+++ src/java-apps/news/src/main/java/in/zycon/demo/gateway/partner/model/fox/Item.java
@@ -1,0 +1,12 @@
+package in.zycon.demo.gateway.partner.model.fox;
+
+import lombok.Data;
+
+@Data
+public class Item {
+    String title;
+    String link;
+    String guid;
+    String description;
+    String pubDate;
+}
