+++ src/java-apps/news/src/main/java/in/zycon/demo/gateway/partner/model/techrepublic/Channel.java
@@ -1,0 +1,17 @@
+package in.zycon.demo.gateway.partner.model.techrepublic;
+
+import com.fasterxml.jackson.dataformat.xml.annotation.JacksonXmlElementWrapper;
+import lombok.Data;
+
+import javax.xml.bind.annotation.XmlElement;
+import java.util.ArrayList;
+import java.util.List;
+
+@Data
+public class Channel {
+    String title;
+    String description;
+    @JacksonXmlElementWrapper(useWrapping = false)
+    @XmlElement
+    List<Item> item = new ArrayList<>();
+}
