--- src/java-apps/news/src/main/java/in/zycon/demo/news/NewsApplication.java
@@ -1,30 +1,0 @@
-package in.zycon.demo.news;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.context.annotation.Bean;
-import springfox.documentation.builders.PathSelectors;
-import springfox.documentation.builders.RequestHandlerSelectors;
-import springfox.documentation.service.ApiInfo;
-import springfox.documentation.spi.DocumentationType;
-import springfox.documentation.spring.web.plugins.Docket;
-import springfox.documentation.swagger2.annotations.EnableSwagger2;
-
-@SpringBootApplication
-@EnableSwagger2
-public class NewsApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(NewsApplication.class, args);
-    }
-
-    @Bean
-    public Docket api() {
-        return new Docket(DocumentationType.SWAGGER_2)
-                .select()
-                .apis(RequestHandlerSelectors.basePackage("in.zycon.demo.news.controller"))
-                .paths(PathSelectors.any())
-                .build();
-    }
-
-}
