--- src/java-apps/news/src/main/java/in/zycon/demo/news/exception/NewsNotFound.java
@@ -1,8 +1,0 @@
-package in.zycon.demo.news.exception;
-
-public class NewsNotFound extends RuntimeException {
-
-    public NewsNotFound(String message) {
-        super(message);
-    }
-}
