--- src/java-apps/news/src/main/java/in/zycon/demo/news/exception/PartnerDataException.java
@@ -1,8 +1,0 @@
-package in.zycon.demo.news.exception;
-
-public class PartnerDataException extends RuntimeException {
-
-    public PartnerDataException(String message) {
-        super(message);
-    }
-}
