--- src/java-apps/news/src/main/java/in/zycon/demo/news/partner/FoxSports.java
@@ -1,24 +1,0 @@
-package in.zycon.demo.news.partner;
-
-import in.zycon.demo.news.partner.model.fox.Rss;
-import in.zycon.demo.news.properties.NewsPartner;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.ResponseEntity;
-import org.springframework.stereotype.Component;
-import org.springframework.web.client.RestTemplate;
-
-@Component
-public class FoxSports {
-    @Autowired
-    Helper helper;
-
-    NewsPartner newsPartner;
-
-    public Rss getResponse() {
-        newsPartner = helper.getPartnerByName("fox");
-        RestTemplate restTemplate = new RestTemplate();
-        return restTemplate.getForEntity(newsPartner.getUrl(), Rss.class).getBody();
-    }
-
-
-}
