--- src/java-apps/news/src/main/java/in/zycon/demo/news/partner/model/fox/Item.java
@@ -1,12 +1,0 @@
-package in.zycon.demo.news.partner.model.fox;
-
-import lombok.Data;
-
-@Data
-public class Item {
-    String title;
-    String link;
-    String guid;
-    String description;
-    String pubDate;
-}
