--- src/java-apps/news/src/main/java/in/zycon/demo/news/partner/model/techrepublic/Item.java
@@ -1,16 +1,0 @@
-package in.zycon.demo.news.partner.model.techrepublic;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-public class Item {
-    String title;
-    String description;
-    String pubDate;
-    String link;
-}
