--- src/java-apps/news/src/main/java/in/zycon/demo/news/partner/model/techrepublic/Rss.java
@@ -1,13 +1,0 @@
-package in.zycon.demo.news.partner.model.techrepublic;
-
-import lombok.Data;
-
-import javax.xml.bind.annotation.XmlElement;
-import javax.xml.bind.annotation.XmlRootElement;
-
-@XmlRootElement
-@Data
-public class Rss {
-    @XmlElement
-    Channel channel;
-}
