--- src/java-apps/news/src/main/java/in/zycon/demo/news/properties/NewsPartner.java
@@ -1,17 +1,0 @@
-package in.zycon.demo.news.properties;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-public class NewsPartner {
-    Integer id;
-    String name;
-    String url;
-    String type;
-    Boolean enabled;
-
-}
