+++ src/java-apps/news/src/test/java/in/zycon/demo/gateway/GatewayApplicationTests.java
@@ -1,0 +1,13 @@
+package in.zycon.demo.gateway;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class GatewayApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
