--- src/java-apps/news/src/test/java/in/zycon/demo/news/GatewayApplicationTests.java
@@ -1,13 +1,0 @@
-package in.zycon.demo.news;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class GatewayApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
