+++ src/java-apps/pom.xml
@@ -15,6 +15,7 @@
     <name>microservices</name>
     <modules>
         <module>news</module>
+        <module>gateway</module>
     </modules>
     <properties>
         <java.version>1.8</java.version>
